pyproject.toml
[project]
name = "my-python-project"
version = "0.1.0"
requires-python = ">= 3.11.13"
dependencies = [  
    "streamlit",
    "gspread",
    "oauth2client",
    "google-auth",
    "google-auth-oauthlib",
    "google-auth-httplib2",
    "pandas"]
